main_page divided on model and widget
